data-collection/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
lxml>=4.9.0
webdriver-manager>=4.0.0

# PostgreSQL support
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Google Drive API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Progress bars and utilities
tqdm>=4.66.0
colorlog>=6.8.0

# Configuration
PyYAML>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Cloud Provider SDKs (for multi-cloud VM deployment)
python-digitalocean>=1.17.0
vultr>=2.1.0
linode-cli>=5.48.0
oci>=2.119.1

# SSH and System
paramiko>=3.4.0
fabric>=3.2.0

# Web Framework (for monitoring dashboard)
Flask>=3.0.0
Flask-SocketIO>=5.3.5
